fix(arc-1937): render <br> tags as new lines in object descriptions
https://meemoo.atlassian.net/browse/ARC-1937
